+++ config/overlays/CHECKPS.BIN.yaml
@@ -42,8 +42,10 @@ segments:
     subsegments:
       - [0x1, rodata, rodata]
       - [0xA5, c, code]
+      - [0xE95, c, code4]
       - [0xEA5, c, unk1]
-      - [0x19B1, c, code3]
+      - [0x19B1, c, code5]
+      - [0x1BC1, c, code3]
       - [0x1F45, c, code2]
       - [0x27B9, rodata, table]
   - [0xA77D]
