+++ src/overlays/checkps/code.c
@@ -664,10 +664,3 @@ void UpdateControllerInput(void)
     D_800610A4 = finalButtonState;
     D_800610A8 = 0xF;
 }
-
-/**
- * decomp.me link (95%) https://decomp.me/scratch/EuGt8
- */
-void func_80050B04(void) {
-    D_8005CFE8 = 1;
-}
