+++ src/overlays/checkps/code4.c
@@ -1,0 +1,8 @@
+#include "checkps.h"
+
+/**
+ * decomp.me link (95%) https://decomp.me/scratch/EuGt8
+ */
+void func_80050B04(void) {
+    D_8005CFE8 = 1;
+}
